autorização schoolmember

+++ app/Http/Controllers/SchoolMemberController.php
@@ -19,6 +19,7 @@ class SchoolMemberController extends Controller
 
     public function store(SchoolMemberRequest $request)
     {
+        $this->authorize('create', SchoolMember::class);
         $validated = $request->validated();
         $id = $validated['grade_class_id'];
         SchoolMember::create($validated);
@@ -27,13 +28,17 @@ class SchoolMemberController extends Controller
 
     public function show($id)
     {
+
         $student = SchoolMember::findOrFail($id);
+        $this->authorize('view', $student);
         return view('school_member.show', compact('student'));
     }
 
     public function update(SchoolMemberRequest $request, $id)
     {
+
         $schoolMember = SchoolMember::findOrFail($id);
+        $this->authorize('update', $schoolMember);
         $validated = $request->validated();
         $schoolMember->update($validated);
         return redirect("/grade_class/$schoolMember->grade_class_id/students");
@@ -42,6 +47,8 @@ class SchoolMemberController extends Controller
     public function destroy($id)
     {
         $schoolMember = SchoolMember::findOrFail($id);
+
+        $this->authorize('delete', $schoolMember);
         $class = $schoolMember->grade_class_id;
         $schoolMember->delete();
         return redirect("/grade_class/$class/students");
@@ -58,6 +65,7 @@ class SchoolMemberController extends Controller
 
     public function answeredTests($id) {
         $student = SchoolMember::findOrFail($id);
+        $this->authorize('viewAnsTests', $student);
         return $student->answeredTests;
     }
 
@@ -70,6 +78,7 @@ class SchoolMemberController extends Controller
 
     public function getTopicScore($student_id, $test_id, $topic_id) {
         $student = SchoolMember::findOrFail($student_id);
+        $this->authorize('viewTopicScore', $student);
         $score = (new TestController)->topicCount($test_id, $topic_id, $student_id);
         return $score;
     }
@@ -77,6 +86,7 @@ class SchoolMemberController extends Controller
 
     public function getMemberbyEnroll($enroll) {
         $member = SchoolMember::where('enroll', $enroll) ->get();
+        $this->authorize('view', $member);
         return $member;
     }
